# User/Role/Feature lookup improvements