Requires https://github.com/ManageIQ/manageiq/pull/20506

Makes use of new scopes to improve the number of queries required for login and even regular requests (though, the latter is less significant).

Also avoids some duplicate `User` lookups on login.


Benchmarks
----------

Note:  Improvements are in number of **queries**.


### Before

```
$ bundle exec miqperf benchmark -ac2 "/api/users"
$ bundle exec miqperf report --last
/api/auth
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 5088 |      26 |        183 |   10 |
/api/users
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
|  267 |      12 |         13 | 1494 |
|   17 |      10 |        3.9 |   12 |
```


### After

```
$ bundle exec miqperf benchmark -ac2 "/api/users"
$ bundle exec miqperf report --last
/api/auth
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 4949 |      12 |        181 |    1 |
/api/users
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
|  258 |       7 |         12 | 1489 |
|   15 |       5 |        3.3 |    7 |
```


Links
-----

* https://github.com/ManageIQ/manageiq/pull/20506
* More improvements as part of this effort:  https://github.com/ManageIQ/manageiq-api/issues/880